lots of fixes for android translation and more

+++ ZColor.swift
@@ -5,29 +5,29 @@
 //  Created by Tor Langballe on /23/9/14.
 //  Copyright (c) 2014 Capsule.fm. All rights reserved.
 //
-// #package com.github.torlangballe.zetrus
+// #package com.github.torlangballe.CetrusAndroid
 
 struct ZHSBA {
-    var h:Float = Float(0.0)
-    var s:Float = Float(0.0)
-    var b:Float = Float(0.0)
-    var a:Float = Float(0.0)
+    var h:Double = 0.0
+    var s:Double = 0.0
+    var b:Double = 0.0
+    var a:Double = 0.0
 }
 
 struct ZRGBA {
-    var r:Float = Float(0.0)
-    var g:Float = Float(0.0)
-    var b:Float = Float(0.0)
-    var a:Float = Float(0.0)
+    var r:Double = 0.0
+    var g:Double = 0.0
+    var b:Double = 0.0
+    var a:Double = 0.0
 }
 
 extension ZColor {
-    func OpacityChanged(_ opacity:Float32) -> ZColor {
+    func OpacityChanged(_ opacity:Double) -> ZColor {
         let c = RGBA
         return ZColor(r:c.r, g:c.g, b:c.b, a:opacity)
     }
     
-    func Mix(_ withColor: ZColor, amount: Float32) -> ZColor {
+    func Mix(_ withColor: ZColor, amount:Double) -> ZColor {
         let wc = withColor.RGBA
         var c = RGBA
         c.r = (1 - amount) * c.r + wc.r * amount
@@ -37,13 +37,13 @@ extension ZColor {
         return ZColor(r:c.r, g:c.g, b:c.b, a:c.a)
     }
     
-    func MultipliedBrightness(_ multiply: Float32) -> ZColor {
+    func MultipliedBrightness(_ multiply:Double) -> ZColor {
         let hsba = self.HSBA
         return ZColor(h:hsba.h, s:hsba.s, b:hsba.b * multiply, a:hsba.a)
     }
     
-    func AlteredContrast(_ contrast:Float) -> ZColor {
-        let multi = Float(ZMath.Pow((Double(1.0 + contrast)) / 1.0, 2.0))
+    func AlteredContrast(_ contrast:Double) -> ZColor {
+        let multi = ZMath.Pow((Double(1.0 + contrast)) / 1.0, 2.0)
         var c = self.RGBA
         c.r = (c.r - 0.5) * multi + 0.5
         c.g = (c.g - 0.5) * multi + 0.5
