+++ ZColorIOS.swift
@@ -3,7 +3,7 @@
 //
 //  Created by Tor Langballe on /13/7/18.
 //
-// #package com.github.torlangballe.zetrus
+// #package com.github.torlangballe.CetrusAndroid
 
 import Foundation
 
@@ -24,15 +24,15 @@ class ZColor : Hashable {
         undefined = false
         self.color = color
     }
-    init(white:Float32, a:Float32 = 1.0) {
+    init(white:Double, a:Double = 1.0) {
         undefined = false
         color = UIColor(white:CGFloat(white), alpha:CGFloat(a))
     }
-    init(r:Float32, g:Float32, b:Float32, a:Float32 = 1.0) {
+    init(r:Double, g:Double, b:Double, a:Double = 1.0) {
         undefined = false
         color = UIColor(red:CGFloat(r), green:CGFloat(g), blue:CGFloat(b), alpha:CGFloat(a))
     }
-    init(h: Float32, s: Float32, b: Float32, a: Float32 = 1.0) {
+    init(h: Double, s: Double, b: Double, a: Double = 1.0) {
         undefined = false
         color = UIColor(hue:CGFloat(h), saturation:CGFloat(s), brightness:CGFloat(b), alpha:CGFloat(a))
     }
@@ -41,10 +41,10 @@ class ZColor : Hashable {
         (h, s, b, a) = (0, 0, 0, 0)
         color.getHue(&h, saturation:&s, brightness:&b, alpha:&a)
         var hsba = ZHSBA()
-        hsba.h = Float(h)
-        hsba.s = Float(h)
-        hsba.b = Float(h)
-        hsba.a = Float(h)
+        hsba.h = Double(h)
+        hsba.s = Double(h)
+        hsba.b = Double(h)
+        hsba.a = Double(h)
         return hsba
     }
     
@@ -58,30 +58,30 @@ class ZColor : Hashable {
         var c = ZRGBA()
         (r, g, b, a) = (0, 0, 0, 0)
         color.getRed(&r, green:&g, blue:&b, alpha:&a)
-        c.r = min(Float(r), 1)
-        c.g = min(Float(g), 1)
-        c.b = min(Float(b), 1)
-        c.a = min(Float(a), 1)
+        c.r = Double(min(r, 1))
+        c.g = Double(min(g, 1))
+        c.b = Double(min(b, 1))
+        c.a = Double(min(a, 1))
         return c
     }
     
-    var GrayScaleAndAlpha: (Float, Float) { // white, alpha
+    var GrayScaleAndAlpha: (Double, Double) { // white, alpha
         var w, a: CGFloat
         (w, a) = (0, 0)
         color.getWhite(&w, alpha:&a)
-        return (Float(w), Float(a))
+        return (Double(w), Double(a))
     }
-    var GrayScale: Float {
+    var GrayScale: Double {
         var w:CGFloat = 0
         var a:CGFloat = 0
         color.getWhite(&w, alpha:&a)
-        return Float(w)
+        return Double(w)
     }
-    var Opacity: Float {
+    var Opacity: Double {
         var w, a: CGFloat
         (w, a) = (0, 0)
         color.getWhite(&w, alpha:&a)
-        return Float(a)
+        return Double(a)
     }
 
     var rawColor: UIColor { return color }
